refactor tests into multiple files
